requirements.txt
requests
selenium
bs4
colorama
pycryptodome
pypiwin32
PIL-Tools
pylibcheck
pyinstaller
psutil
keyboard
httpx
